Incorporacion de foreign keys y plantilla basica para poder comprobar navegacion

+++ src/main/java/com/proyecto_gym/repository/PreguntaFrecuenteRepository.java
@@ -1,9 +1,7 @@
 package com.proyecto_gym.repository;
 
-
 import com.proyecto_gym.domain.PreguntaFrecuente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PreguntaFrecuenteRepository extends JpaRepository<PreguntaFrecuente, Long> {
-
+public interface PreguntaFrecuenteRepository extends JpaRepository<PreguntaFrecuente, Integer> {
 }
